refactor(ui): unify working spinner into one shared WorkingIndicator
The Mac sidebar and iOS both drew an 'agent is working' comet, but from two
separate copies that had drifted: the Mac copy had a polished two-signal render
(brightness wave + head size swell, 2.5pt dots, 0.5 tail floor) while the iOS
copy was still opacity-only with a near-invisible tail.

Fold Mac's polished version into the shared TermioShared.WorkingIndicator (now
carrying the optional phase parameter the menu bar needs) and delete the Mac
duplicate, so both platforms render from one implementation and can't drift
again. Tint stays caller-injected: the Mac sidebar and the iOS list both pass
adaptive monochrome ink (color is reserved for the done/attention status dots),
while the menu bar keeps its own black/white tint.
