+++ Shared/Sources/TermioShared/SessionStatus.swift
@@ -53,12 +53,21 @@ public struct StatusDot: View {
 /// The "agent is working" mark: a 3×3 grid of dots with a bright comet that
 /// orbits the eight perimeter cells, so the small nine-square grid reads as
 /// rotating. Sits in place of the session's brand icon while a turn is in
-/// flight, tinted with the agent's brand color.
+/// flight. The caller supplies the tint — the agent's brand color on iOS, and
+/// monochrome ink in the Mac sidebar, where the vibrancy material would wash a
+/// brand tint to grey.
 public struct WorkingIndicator: View {
     let tint: Color
 
-    public init(tint: Color = .secondary) {
+    /// When nil the comet self-animates via `TimelineView`; supplying a phase
+    /// (0...1) renders one still frame instead, so a caller can drive the
+    /// rotation with its own timer where `TimelineView` never ticks — e.g.
+    /// inside an open `NSMenu`, which runs a modal event-tracking loop.
+    let phase: Double?
+
+    public init(tint: Color = .secondary, phase: Double? = nil) {
         self.tint = tint
+        self.phase = phase
     }
 
     /// The eight perimeter cells of the 3×3 grid in clockwise order, as
@@ -66,43 +75,70 @@ public struct WorkingIndicator: View {
     private static let ring: [(Int, Int)] = [
         (0, 0), (1, 0), (2, 0), (2, 1), (2, 2), (1, 2), (0, 2), (0, 1),
     ]
-    private let dotSize: CGFloat = 2.3
+    // Dots this small read lighter than their nominal opacity, so the size and
+    // the opacity ramp are tuned together: 2.5pt with a 0.5 tail floor sits at
+    // the same perceived weight as the neighboring 15pt glyphs.
+    private let dotSize: CGFloat = 2.5
     private let spacing: CGFloat = 3.6
     private let period: Double = 1.1
 
     public var body: some View {
-        TimelineView(.animation) { context in
-            let phase = context.date.timeIntervalSinceReferenceDate
-                .truncatingRemainder(dividingBy: period) / period
-            ZStack {
-                // A faint steady center anchors the spinning ring.
-                dot(opacity: 0.3)
-                ForEach(Array(Self.ring.enumerated()), id: \.offset) { index, cell in
-                    dot(opacity: opacity(at: index, phase: phase))
-                        .offset(
-                            x: CGFloat(cell.0 - 1) * spacing,
-                            y: CGFloat(cell.1 - 1) * spacing
-                        )
-                }
+        if let phase {
+            grid(phase: phase)
+        } else {
+            TimelineView(.animation) { context in
+                let p = context.date.timeIntervalSinceReferenceDate
+                    .truncatingRemainder(dividingBy: period) / period
+                grid(phase: p)
             }
-            .frame(width: 13, height: 13)
         }
     }
 
-    private func dot(opacity: Double) -> some View {
+    private func grid(phase: Double) -> some View {
+        ZStack {
+            // A steady center anchors the spinning ring, matching the tail so
+            // the grid reads as one solid mark with a swell running around it.
+            dot(opacity: 0.5, scale: 1)
+            ForEach(Array(Self.ring.enumerated()), id: \.offset) { index, cell in
+                let distance = ringDistance(at: index, phase: phase)
+                dot(opacity: opacity(distance: distance), scale: scale(distance: distance))
+                    .offset(
+                        x: CGFloat(cell.0 - 1) * spacing,
+                        y: CGFloat(cell.1 - 1) * spacing
+                    )
+            }
+        }
+        .frame(width: 13, height: 13)
+    }
+
+    private func dot(opacity: Double, scale: Double) -> some View {
         Circle()
             .fill(tint)
-            .frame(width: dotSize, height: dotSize)
+            .frame(width: dotSize * scale, height: dotSize * scale)
             .opacity(opacity)
     }
 
-    /// Brightness of a perimeter cell: peaks at the comet's head and fades over
-    /// the next few cells, measured the shorter way around so the tail wraps.
-    private func opacity(at index: Int, phase: Double) -> Double {
+    /// A perimeter cell's distance from the comet's head, measured the shorter
+    /// way around the ring so the tail wraps.
+    private func ringDistance(at index: Int, phase: Double) -> Double {
         let count = Double(Self.ring.count)
         let head = phase * count
         let raw = abs(Double(index) - head)
-        let distance = min(raw, count - raw)
-        return max(0.22, 1 - distance / 3)
+        return min(raw, count - raw)
+    }
+
+    /// The rotation is carried by two signals so neither has to be extreme: a
+    /// brightness wave AND a size swell at the comet's head. Opacity alone
+    /// needed a near-invisible tail to read as motion, which left the whole mark
+    /// far paler than the full-ink glyphs beside it; with the swell doing half
+    /// the work the tail floor stays at half ink and the grid keeps real weight.
+    private func opacity(distance: Double) -> Double {
+        max(0.5, 1 - distance / 4)
+    }
+
+    /// Size factor for a cell: the head swells a fifth and the swell dies out
+    /// over the next two cells.
+    private func scale(distance: Double) -> Double {
+        1 + 0.2 * max(0, 1 - distance / 2)
     }
 }
